Cleanup, make sure NetexMapper is the single point of entry, the aggregate root of mapping. Make sub-mapper classes package local. Move unrelated classes out of the package.
