+++ src/test/java/org/opentripplanner/netex/support/DayTypeRefToServiceIdMapperTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.netex.mapping;
+package org.opentripplanner.netex.support;
 
 import org.junit.Test;
 
@@ -7,7 +7,7 @@ import java.util.Collections;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
-import static org.opentripplanner.netex.mapping.DayTypeRefToServiceIdMapper.generateServiceId;
+import static org.opentripplanner.netex.support.DayTypeRefToServiceIdMapper.generateServiceId;
 
 /**
  * @author Thomas Gran (Capra) - tgr@capraconsulting.no (01.12.2017)
